requirements.txt
openai
duckduckgo-search
pyttsx3
speechrecognition
pyaudio
gtts
translate
playsound
keyboard
python-dotenv